tech(api): move admin/frameworks/* routes to prescription target profile bounded context and use framework api
