--- api/lib/application/frameworks/frameworks-controller.js
@@ -1,26 +1,0 @@
-import { extractLocaleFromRequest } from '../../../src/shared/infrastructure/utils/request-response-utils.js';
-import { usecases } from '../../domain/usecases/index.js';
-import * as frameworkAreasSerializer from '../../infrastructure/serializers/jsonapi/framework-areas-serializer.js';
-
-const getFrameworkAreas = async function (request, h, dependencies = { frameworkAreasSerializer }) {
-  const frameworkId = request.params.id;
-  const areas = await usecases.getFrameworkAreas({ frameworkId });
-  return dependencies.frameworkAreasSerializer.serialize(areas);
-};
-
-const getPixFrameworkAreasWithoutThematics = async function (
-  request,
-  h,
-  dependencies = { extractLocaleFromRequest, frameworkAreasSerializer },
-) {
-  const locale = dependencies.extractLocaleFromRequest(request);
-  const areas = await usecases.getFrameworkAreas({ frameworkName: 'Pix', locale });
-  return dependencies.frameworkAreasSerializer.serialize(areas, { withoutThematics: true });
-};
-
-const frameworksController = {
-  getFrameworkAreas,
-  getPixFrameworkAreasWithoutThematics,
-};
-
-export { frameworksController };
