+++ api/lib/routes.js
@@ -1,10 +1,9 @@
 import * as healthcheck from '../src/shared/application/healthcheck/index.js';
 import * as campaignParticipations from './application/campaign-participations/index.js';
-import * as frameworks from './application/frameworks/index.js';
 import * as organizations from './application/organizations/index.js';
 import * as scoOrganizationLearners from './application/sco-organization-learners/index.js';
 import * as users from './application/users/index.js';
 
-const routes = [campaignParticipations, healthcheck, organizations, scoOrganizationLearners, frameworks, users];
+const routes = [campaignParticipations, healthcheck, organizations, scoOrganizationLearners, users];
 
 export { routes };
