+++ api/src/prescription/target-profile/domain/usecases/find-areas-for-framework-name.js
@@ -1,0 +1,43 @@
+import { AreaForAdmin } from '../../../../shared/domain/models/index.js';
+
+export async function findAreasForFrameworkName({
+  frameworkName,
+  locale,
+  skillRepository,
+  tubeRepository,
+  thematicRepository,
+  areaRepository,
+  frameworkRepository,
+}) {
+  const framework = await frameworkRepository.getByName({ name: frameworkName });
+  const areasWithCompetences = await areaRepository.findByFrameworkIdWithCompetences({
+    frameworkId: framework.id,
+    locale,
+  });
+
+  const competences = areasWithCompetences.flatMap((area) => area.competences);
+
+  const competenceIds = competences.map(({ id: competenceId }) => competenceId);
+  const thematics = await thematicRepository.findByCompetenceIds(competenceIds, locale);
+
+  const tubeIds = thematics.flatMap((thematic) => thematic.tubeIds);
+  const tubes = await tubeRepository.findActiveByRecordIds(tubeIds, locale);
+
+  const skillIds = tubes.flatMap((tube) => tube.skillIds);
+  const skills = await skillRepository.findActiveByRecordIds(skillIds);
+
+  return areasWithCompetences.map(
+    (areaWithCompetences) =>
+      new AreaForAdmin({
+        id: areaWithCompetences.id,
+        frameworkId: areaWithCompetences.frameworkId,
+        title: areaWithCompetences.title,
+        code: areaWithCompetences.code,
+        color: areaWithCompetences.color,
+        allCompetences: areaWithCompetences.competences,
+        allThematics: thematics,
+        allTubes: tubes,
+        allSkills: skills,
+      }),
+  );
+}
