+++ api/src/prescription/target-profile/infrastructure/serializers/jsonapi/framework-areas-serializer.js
